Extract a stream_ext module that will be used only in tests
This contains a useful `wait_next` method that can be invoked on
streams, which waits for the next item to be resolved using the supplied
reactor::Core without moving the stream.
